Break out mnemonic into its own component
