+++ renderer/components/OnboardingFlow/ImportAccount/ImportAccount.tsx
@@ -1,23 +1,13 @@
-import { Box, Heading, Text } from "@chakra-ui/react";
+import { Box, Heading } from "@chakra-ui/react";
 import Link from "next/link";
 import { useRouter } from "next/router";
-import { useMemo, useState } from "react";
 
 import { trpcReact } from "@/providers/TRPCProvider";
-import {
-  EMPTY_PHRASE_ARRAY,
-  MnemonicPhrase,
-} from "@/ui/Forms/MnemonicPhrase/MnemonicPhrase";
-import { TextInput } from "@/ui/Forms/TextInput/TextInput";
-import { PillButton } from "@/ui/PillButton/PillButton";
-import { validateMnemonic } from "@/utils/mnemonic";
+
+import { MnemonicImport } from "./MnemonicImport";
 
 export function ImportAccount() {
   const router = useRouter();
-  const [accountName, setAccountName] = useState("");
-  const [isAccountNameDirty, setIsAccountNameDirty] = useState(false);
-  const [phraseValues, setPhraseValues] =
-    useState<Array<string>>(EMPTY_PHRASE_ARRAY);
 
   const {
     mutate: importAccount,
@@ -25,62 +15,20 @@ export function ImportAccount() {
     error,
   } = trpcReact.importAccount.useMutation();
 
-  const errorMessage = useMemo(() => {
-    return validateMnemonic(phraseValues);
-  }, [phraseValues]);
-
-  const hasValidName = accountName.length > 0;
-  const hasNameError = isAccountNameDirty && !hasValidName;
-
   return (
     <Box>
       <Link href="/onboarding">Back</Link>
       <Heading mt={24} mb={8}>
         Import Account
       </Heading>
-      <TextInput
-        label="Account Name"
-        value={accountName}
-        error={
-          isAccountNameDirty && !hasValidName
-            ? "Please enter a name for this account"
-            : null
-        }
-        onChange={(e) => {
-          setAccountName(e.target.value);
-        }}
-        onBlur={() => {
-          setIsAccountNameDirty(true);
-        }}
-      />
-      <Text mt={8} mb={4}>
-        Please enter your mnemonic phrase. If you&apos;ve copied the full phrase
-        to your clipboard, you can paste it in any input and it will
-        automatically be split into the correct number of words.
-      </Text>
-      <MnemonicPhrase
-        defaultVisible
-        phrase={phraseValues}
-        error={errorMessage || error?.shape?.message}
-        onChange={(value) => {
-          setPhraseValues(value);
-        }}
-        mb={8}
-      />
-      <PillButton
-        height="60px"
-        px={8}
-        isDisabled={isLoading || !!errorMessage || hasNameError}
-        onClick={() => {
-          if (!hasValidName) {
-            setIsAccountNameDirty(true);
-            return;
-          }
-
+      <MnemonicImport
+        isLoading={isLoading}
+        error={error?.shape?.message}
+        handleImport={({ name, account }) => {
           importAccount(
             {
-              name: accountName,
-              account: phraseValues.join(" "),
+              name,
+              account,
             },
             {
               onSuccess: () => {
@@ -89,9 +37,7 @@ export function ImportAccount() {
             },
           );
         }}
-      >
-        Continue
-      </PillButton>
+      />
     </Box>
   );
 }
